Sum for BinnedStatistic added (pure).
